+++ app/src/main/java/com/giraffe/weatherforecasapplication/model/Daily.kt
@@ -1,23 +1,23 @@
 package com.giraffe.weatherforecasapplication.model
 
 data class Daily(
-    val clouds: Int,
+    val clouds: Double,
     val dew_point: Double,
-    val dt: Int,
+    val dt: Double,
     val feels_like: FeelsLike,
-    val humidity: Int,
+    val humidity: Double,
     val moon_phase: Double,
-    val moonrise: Int,
-    val moonset: Int,
+    val moonrise: Double,
+    val moonset: Double,
     val pop: Double,
-    val pressure: Int,
+    val pressure: Double,
     val rain: Double,
-    val sunrise: Int,
-    val sunset: Int,
+    val sunrise: Double,
+    val sunset: Double,
     val temp: Temp,
     val uvi: Double,
     val weather: List<Weather>,
-    val wind_deg: Int,
+    val wind_deg: Double,
     val wind_gust: Double,
     val wind_speed: Double
 )
